v2

+++ InventoryManagement/Controllers/CustomerController.cs
@@ -1,4 +1,7 @@
-﻿using InventoryManagement.Models;
+﻿using InventoryManagement.Dto.Request;
+using InventoryManagement.Dto.Response;
+using InventoryManagement.Mapper;
+using InventoryManagement.Models;
 using InventoryManagement.Repository;
 using Microsoft.AspNetCore.Mvc;
 
@@ -16,24 +19,41 @@ namespace InventoryManagement.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Customer>>> GetCustomers()
+        public async Task<ActionResult<IEnumerable<CustomerResponse>>> GetCustomers()
         {
+            var customerResponse = new List<CustomerResponse>();
+
             var customers = await _repository.GetAllAsync();
-            return Ok(customers);
+
+            foreach (var item in customers)
+            {
+                customerResponse.Add(CustomerMapper.CustomerMapperView(item));
+            }
+
+            return customerResponse;
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<Customer>> GetCustomer(int id)
+        public async Task<ActionResult<CustomerResponse>> GetCustomer(int id)
         {
+            var customerResponse = new CustomerResponse();
+
             var customer = await _repository.GetByIdAsync(id);
+
             if (customer == null) return NotFound();
-            return Ok(customer);
+
+            customerResponse = CustomerMapper.CustomerMapperView(customer);
+
+            return customerResponse;
         }
 
         [HttpPost]
-        public async Task<ActionResult<Customer>> CreateCustomer(Customer customer)
+        public async Task<ActionResult<Customer>> CreateCustomer(CustomerRequest customerRequest)
         {
+            var customer = CustomerMapper.CustomerMapperDto(customerRequest);
+
             var createdCustomer = await _repository.AddAsync(customer);
+
             return CreatedAtAction(nameof(GetCustomer), new { id = createdCustomer.Id }, createdCustomer);
         }
 
@@ -41,14 +61,21 @@ namespace InventoryManagement.Controllers
         public async Task<IActionResult> UpdateCustomer(int id, Customer customer)
         {
             if (id != customer.Id) return BadRequest();
+
             await _repository.UpdateAsync(customer);
+
             return NoContent();
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteCustomer(int id)
         {
+            var customer = await _repository.GetByIdAsync(id);
+            
+            if (customer == null) return NotFound();
+
             await _repository.DeleteAsync(id);
+
             return NoContent();
         }
     }
